minus angle done
