+++ GccApplication1.atsln
@@ -9,6 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	ProjectSection(SolutionItems) = preProject
 		GccApplication1\AngleRawData.h = GccApplication1\AngleRawData.h
 		GccApplication1\I2C.h = GccApplication1\I2C.h
+		GccApplication1\Interrupt.h = GccApplication1\Interrupt.h
 		GccApplication1\LCD.h = GccApplication1\LCD.h
 		GccApplication1\MPU6050.h = GccApplication1\MPU6050.h
 		GccApplication1\USART.h = GccApplication1\USART.h
